编解码出错

+++ gameserver/src/main/java/com/game/gameserver/annotation/CmdHandler.java
@@ -12,5 +12,5 @@ import java.lang.annotation.Target;
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
 public @interface CmdHandler {
-    int cmd() default 0;
+    short cmd() default 0;
 }
